make PeriodicUpdate work without needing the primary index, update backends and frontends to select job type, update history graph to only show CensusScan
