+++ conf/routes
@@ -12,7 +12,7 @@ GET     /api/stats/unattached       @controllers.StatsController.unattachedStats
 
 GET     /api/nearline/currentState  @controllers.NearlineDataController.currentStateData
 
-GET     /api/jobs/forTimespan           @controllers.JobsController.jobsForTimespan(startAt:Option[String]?=None,endAt:Option[String]?=None,showRunning:Boolean?=false)
+GET     /api/jobs/:jobType/forTimespan  @controllers.JobsController.jobsForTimespan(jobType, startAt:Option[String]?=None,endAt:Option[String]?=None,showRunning:Boolean?=false)
 GET     /api/jobs/running               @controllers.JobsController.runningJobs
 GET     /api/jobs/:jobType/lastSuccess  @controllers.JobsController.lastSuccessfulJob(jobType)
